Fix filenames of capture related traits.
