+++ OpenRA.Mods.Common/OpenRA.Mods.Common.csproj
@@ -311,9 +311,9 @@
     <Compile Include="Traits\Buildings\RepairableBuilding.cs" />
     <Compile Include="Traits\Buildings\RequiresBuildableArea.cs" />
     <Compile Include="Traits\Buildings\Reservable.cs" />
-    <Compile Include="Traits\CapturableProgressBar.cs" />
     <Compile Include="Traits\CaptureProgressBar.cs" />
-    <Compile Include="Traits\CaptureProgressBlink.cs" />
+    <Compile Include="Traits\CapturableProgressBar.cs" />
+    <Compile Include="Traits\CapturableProgressBlink.cs" />
     <Compile Include="Traits\Conditions\GrantConditionOnPlayerResources.cs" />
     <Compile Include="Traits\GrantExternalConditionToProduced.cs" />
     <Compile Include="Traits\Multipliers\CreatesShroudMultiplier.cs" />
