+++ OpenRA.Mods.Common/Traits/CapturableProgressBlink.cs
